+++ system/apps/checks.mjs
@@ -1,6 +1,9 @@
 import { AOVactorDetails } from "./actor-details.mjs";
 import AOVDialog from "../setup/aov-dialog.mjs";
 import { AOVSelectLists } from "./select-lists.mjs";
+import { RECard } from "../chat/resistance-chat.mjs";
+import { OPCard } from "../chat/opposed-chat.mjs";
+import { AUCard } from "../chat/augment-chat.mjs";
 
 export class RollType {
   static CHARACTERISTIC = "CH";
@@ -11,6 +14,10 @@ export class RollType {
 
 export class CardType {
   static UNOPPOSED = "NO";
+  static FIXED = "FI";  //Resistance roll against Fixed target (e.g. a poison)
+  static RESISTANCE = "RE";
+  static OPPOSED = "OP";
+  static AUGMENT = "AU";
 }
 
 export class RollResult {
@@ -67,6 +74,7 @@ export class AOVCheck {
       dialogTemplate: "systems/aov/templates/dialog/rollOptions.hbs",
       chatTemplate: "systems/aov/templates/chat/roll-result.hbs",
       state: options.state ?? "open",
+      wait: false,
       particName,
       particId,
       particType,
@@ -78,11 +86,15 @@ export class AOVCheck {
       damTypeLabel: "",
       dbLabel: "s",
       successLevel: "2",
+      successLabel: "",
       db : "",
       difficulty: "",
       diffLabel: "",
       targetScore: 0,
       rawScore: 0,
+      oppRes: 0,
+      targetAdj: 0,
+      augAdj: 0,
       flatMod: options.flatMod ?? 0,
       rollFormula: options.rollFormula ?? "1D100",
       resultLevel: options.resultLevel ?? 0,
@@ -90,7 +102,8 @@ export class AOVCheck {
       userID: game.user._id,                     
     }
       
-    //Adjust config based on roll type
+    //Adjust config based on ROLL TYPE
+    //
     let tempItem = ""
     switch (options.rollType) {
       case RollType.CHARACTERISTIC:
@@ -111,8 +124,6 @@ export class AOVCheck {
             config.rawScore = config.rawScore + particActor.system[tempItem.system.category] ?? 0
           }
         }
-
-
         break;
       case RollType.DAMAGE:
         tempItem = await particActor.items.get(config.skillId)
@@ -123,9 +134,9 @@ export class AOVCheck {
         config.dbLabel = tempItem.system.dbLabel
         config.db = particActor.system.dmgBonus
         //Adjust Damage Bonus for Weapon Type
-        if (tempItem.system.weaponType === "thrown") {
+        if (tempItem.system.damMod === "h") {
           config.db = particActor.system.dmgBonus + "/2"  
-        } else if (tempItem.system.weaponType === "missile") {
+        } else if (tempItem.system.damMod === "n") {
           config.db = "0"
         }
         break;
@@ -139,13 +150,35 @@ export class AOVCheck {
     config.diffLabel = game.i18n.localize("AOV.rolls.simple")
     config.targetScore = config.rawScore
 
-    //Adjust config based on Card Type
+    //Adjust config based on CARD TYPE
+    //
     switch (options.cardType) {
       case CardType.UNOPPOSED:
         config.state = "closed";
         config.chatTemplate =
           "systems/aov/templates/chat/roll-result.hbs";
         break;
+      case CardType.FIXED:
+        config.state = "closed";
+        config.chatTemplate =
+          "systems/aov/templates/chat/roll-result.hbs";
+        config.shiftKey = false;  //always need to open dialog to get target  
+        break;    
+      case CardType.RESISTANCE:
+        config.wait = true;
+        config.chatTemplate =
+          "systems/aov/templates/chat/roll-resistance.hbs";
+        break;    
+      case CardType.OPPOSED:
+        config.wait = true;
+        config.chatTemplate =
+          "systems/aov/templates/chat/roll-opposed.hbs";
+        break;
+      case CardType.AUGMENT:
+        config.wait = true;
+        config.chatTemplate =
+          "systems/aov/templates/chat/roll-augment.hbs";
+        break;        
       default:
         ui.notifications.error(
           game.i18n.format("AOV.ErrorMsg.cardInvalid", {type: options.cardType})
@@ -167,15 +200,18 @@ export class AOVCheck {
     if (!config.shiftKey) {
       let usage = await AOVCheck.RollDialog(config);
       if (usage) {
-         config.flatMod = Number(usage.checkBonus);
-         config.difficulty = usage.diffOption ?? "";
-         config.successLevel = usage.dmgLevel ?? "2";
-         if (usage.ctOption) {
+        config.flatMod = Number(usage.checkBonus);
+        config.difficulty = usage.diffOption ?? "";
+        config.successLevel = usage.dmgLevel ?? "2";
+        if (usage.ctOption) {
           config.damType = usage.ctOption
           config.damTypeLabel = game.i18n.localize('AOV.DamType.'+config.damType)
-         }
-         }
-      }      
+        }
+        config.oppRes = usage.oppRes
+      } else {
+        return
+      }
+    }      
 
     //Where difficulty used
     if (config.difficulty !="") {
@@ -202,6 +238,18 @@ export class AOVCheck {
     //Adjust target Score for check Bonus, reflexive Modifier and calculate critBonus where target score > 20 or <0
     config.targetScore = config.targetScore + config.flatMod
 
+
+    //Adjust Target for Fixed Resistance Roll
+    if (config.cardType === CardType.FIXED) {
+      config.targetScore = config.targetScore - (5*config.oppRes) + 50
+      config.targetScore = Math.max(Math.min(config.targetScore, 95),5)  
+      config.rawScore = config.rawScore/5;
+    }
+
+    //Adjust for Resistance Roll
+    if (config.cardType === CardType.RESISTANCE) {
+      config.rawScore = config.rawScore/5;
+    }
 
     //If damage type is still Cut and Thrust default to Impale
     if (config.damType === 'ct') {
@@ -236,9 +284,10 @@ export class AOVCheck {
     config.successLevelLabel = game.i18n.localize('AOV.resultLevel.'+ config.successLevel)
 
 
-    //Make the Dice Roll
-    await AOVCheck.makeRoll(config);
-
+    //Make the Dice Roll - unless we are told to wait
+    if (!config.wait) {
+      await AOVCheck.makeRoll(config);
+    }  
 
     //Format the data so it's in the same format as will be held in the Chat Message when saved
     let chatMsgData = {
@@ -247,6 +296,7 @@ export class AOVCheck {
       chatType: config.chatType,
       chatTemplate: config.chatTemplate,
       state: config.state,
+      wait: config.wait,
       rolls: config.roll,
       resultLevel: config.resultLevel,
       rollResult: config.rollResult,
@@ -266,11 +316,15 @@ export class AOVCheck {
           diffLabel: game.i18n.localize("AOV.rolls." + config.difficulty),
           rollFormula: config.rollFormula,
           flatMod: config.flatMod,
+          targetAdj: config.targetAdj,
           rollResult: config.rollResult,          
           rollVal: config.rollVal,
           roll: config.roll,
+          oppRes: config.oppRes,
           damTypeLabel: config.damTypeLabel,
+          successLevel: config.successLevel,
           successLevelLabel: config.successLevelLabel,
+          augAdj: config.augAdj,
           diceRolled: config.diceRolled,
           resultLevel: config.resultLevel,
           resultLabel: game.i18n.localize(
@@ -280,6 +334,19 @@ export class AOVCheck {
         },
       ],
     };
+
+    //In some circumstances add details to an existing chat message rather than create a new one
+    if (['RE','OP','AU'].includes(config.cardType)) {
+      //Check to see if there is an open card within paramaters
+      let checkMsgId = await AOVCheck.checkNewMsg(chatMsgData)
+      if (checkMsgId != false) {
+        // Trigger adding check to the card and then stop
+        await RECard.REAdd(chatMsgData, checkMsgId)
+        return
+      }
+    }
+
+
 
     //Create the ChatMessage and Roll Dice
     const html = await AOVCheck.startChat(chatMsgData);
@@ -351,6 +418,7 @@ export class AOVCheck {
   static async startChat(chatMsgData) {
     let html = await foundry.applications.handlebars.renderTemplate(chatMsgData.chatTemplate, chatMsgData);
     return html;
+
   }
 
   // Display the chat card and roll the dice
@@ -383,7 +451,8 @@ export class AOVCheck {
         alias: alias,
       },
     };
-    if (["NO"].includes(chatMsgData.cardType)) {
+    //List the Card Types that will make Dice So Nice show dice rolls
+    if (["NO","FI"].includes(chatMsgData.cardType)) {
       chatData.rolls = [chatMsgData.rolls];
     }
     let msg = await ChatMessage.create(chatData);
@@ -395,14 +464,41 @@ export class AOVCheck {
     let diffOptions = await AOVSelectLists.difficultyOptions();
     let dmgLevels = await AOVSelectLists.dmgLevels();    
     let ctOptions = await AOVSelectLists.cutThrust();
+    let cardLabel = game.i18n.localize('AOV.card.'+options.cardType)
+    let askFixed = false
+    let askDiff = false
+    let askSuccess = false
+    let askDamType = false
+    let askBonus = true
+    if (options.rollType === 'DM') {
+      cardLabel = game.i18n.localize('AOV.rolls.DM')
+      askBonus = false
+      askSuccess = true
+      if (options.damType === 'ct') {
+        askDamType = true
+      }
+    }
+    if (['FI'].includes(options.cardType)) {
+      askFixed = true
+    }
+    if (options.cardType === 'NO' && options.rollType === 'CH') {
+      askDiff = true
+    }
+
 
     const data = {
       cardType: options.cardType,
+      cardLabel,
       label: options.label,
       rollType: options.rollType,
       flatMod: options.flatMod,
       damType: options.damType,
       dmgLevels,
+      askFixed,
+      askDiff,
+      askSuccess,
+      askDamType,
+      askBonus,
       diffOptions,
       ctOptions,
     };
@@ -419,5 +515,89 @@ export class AOVCheck {
     return dlg
   }
 
+  //Check to see if there is an open card that matches the cardTyoe that's not more than a day old
+  static async checkNewMsg(config) {
+    let messages = ui.chat.collection.filter(message => {
+      if (
+        config.cardType === message.getFlag('aov', 'cardType') &&
+        message.getFlag('aov', 'state') !== 'closed'
+      ) {
+        return true
+      }
+      return false
+    })
 
+    if (messages.length) {
+      // Old messages can't be used if message is more than a day old mark it as resolved
+      const timestamp = new Date(messages[messages.length-1].timestamp)
+      const now = new Date()
+      const timeDiffSec = (now - timestamp) / 1000
+      if (60 * 60 * 24 < timeDiffSec) {
+        await messages[messages.length-1].setFlag('aov', 'state', 'closed')
+        messages = []
+      }
+    }
+
+    if (!messages.length) { return false }
+    else { return messages[messages.length-1].id }
+  }
+
+
+  //Function when Chat Message buttons activated to call socket
+  static async triggerChatButton(event) {
+    const targetElement = event.currentTarget
+    const presetType = targetElement.dataset?.preset
+    const dataset = targetElement.dataset
+    const targetChat = $(targetElement).closest('.message')
+    let targetChatId = targetChat[0].dataset.messageId
+    let origin = game.user.id
+    let originGM = game.user.isGM
+
+    if (game.user.isGM) {
+      AOVCheck.handleChatButton({ presetType, targetChatId, origin, originGM, event, dataset })
+    } /* else {
+      const availableGM = game.users.find(d => d.active && d.isGM)?.id
+      if (availableGM) {
+        game.socket.emit('system.brp', {
+          type: 'chatUpdate',
+          to: availableGM,
+          value: { presetType, targetChatId, origin, originGM, event }
+        })
+      } else {
+        ui.notifications.warn(game.i18n.localize('BRP.noAvailableGM'))
+      }
+    }*/
+  }
+
+
+  //Handle changes to Cards based on the presetType value - will be carried out by a GM
+  static async handleChatButton(data) {
+    const presetType = data.presetType
+    let targetMsg = await game.messages.get(data.targetChatId)
+    switch (presetType) {
+      case "close-card":
+        await RECard.REClose(data)
+        break
+      case "remove-roll":
+        await RECard.RERemove(data)
+        break
+      case "resolve-re-card":
+        await RECard.REResolve(data)
+        break
+      case "resolve-op-card":
+        await OPCard.OPResolve(data)
+        break  
+      case "resolve-au-card":
+        await AUCard.AUResolve(data)
+        break                
+      default:
+        return
+    }
+    const pushhtml = await AOVCheck.startChat(targetMsg.flags.aov)
+    await targetMsg.update({ content: pushhtml })
+
+    return
+  }
+  
+  
 }