+++ system/hooks/render-chat-message.mjs
@@ -1,5 +1,5 @@
 /* global $ */
-import { AOVChat } from '../apps/chat.mjs'
+import { AOVChat } from '../chat/chat.mjs'
 
 export default function (app, html, data) {
   AOVChat.renderMessageHook(app, html, data)