+++ system/item/data/item-npcpower-model.mjs
@@ -6,6 +6,7 @@ export default class AOVNPCPowerModel extends AOVItemBaseModel {
     const fields = foundry.data.fields;
     const requiredInteger = { required: true, nullable: false, integer: true };
     const schema = super.defineSchema();
+    schema.priority = new fields.NumberField({ ...requiredInteger, initial: 0 });  //Priority for displayed order
     return schema
   }