+++ system/item/sheets/weapon-sheet.mjs
@@ -43,6 +43,8 @@ export class AoVWeaponSheet extends AoVItemSheet {
     }
     context.damTypeList = await AOVSelectLists.dmgTypeOptions()
     context.damTypeName = game.i18n.localize('AOV.DamType.' + context.system.damType)
+    context.damModList = await AOVSelectLists.damModOptions()
+    context.damModName = game.i18n.localize('AOV.DamMod.' + context.system.damMod)
     context.equippedOptions = await AOVSelectLists.equippedOptions(this.document.type)
     context.isRange = false
     if (['missile','thrown'].includes(context.system.weaponType)) {
@@ -131,9 +133,31 @@ export class AoVWeaponSheet extends AoVItemSheet {
 
   //Activate event listeners using the prepared sheet HTML
   _onRender(context, _options) {
+    this.element.querySelectorAll('.weapon-type').forEach(n => n.addEventListener("change", this.#weaponType.bind(this)))
   }
 
 
+
+
   //-----------------------ACTIONS-----------------------------------
+
+  //--------------LISTENERS------------------
+  //Handle Change in Weapon Type
+  async #weaponType(event) {
+    event.preventDefault();
+    event.stopImmediatePropagation();
+    let newWpnType = event.target.value
+    let newDamMod = "d"
+    if (newWpnType === 'thrown') {
+      newDamMod = "h"
+    } else if (newWpnType === 'missile') {
+      newDamMod = "n"
+    }
+    await this.item.update({'system.damMod': newDamMod,
+                           'system.weaponType': newWpnType
+    })
+  }
+
+
 
 }